+++ Revo.Infrastructure/Events/Async/PerAggregateAsyncEventSequencer{}.cs
@@ -1,0 +1,10 @@
+﻿namespace Revo.Infrastructure.Events.Async
+{
+    public class PerAggregateAsyncEventSequencer<TListener> : PerAggregateAsyncEventSequencer
+        where TListener : IAsyncEventListener
+    {
+        public PerAggregateAsyncEventSequencer() : base(typeof(TListener).Name)
+        {
+        }
+    }
+}