+++ RELEASE_NOTES.md
@@ -11,6 +11,7 @@
   - CommandExecutionOptions that can change the tenant context to run the command in and other options
 - AsyncEventListenerBindingExtensions for easier async event listener registration
 - convenience methods IReadRepository.GetManyAsync<T> and IReadRepository.FindManyAsync<T> using Guid IDs
+- added PerAggregateAsyncEventSequencer and PerTenantAsyncEventSequencer for common sequencer scenarios
 
 ### Changed
 - breaking change: null tenants now cannot access other tenant's data by default (can be changed by configuration)